Bug 35362: Update patron module pop-up windows with consistent footer markup
This patch updates two patron-related templates in order to style
submission and close buttons in a fixed footer at the bottom of the
window, with markup consistent with other pop-up windows.

The patch also updates the "Update child" popup js so that it uses
the same window-opening function we use elsewhere. I also added form
validation to the category selection form to help prevent errors if
someone submits the form without making a selection.

To test, apply the patch and go to the Patrons module

- Click "New patron" and add a patron with the same first and last name
  of a patron already in the database.
- When you save the new patron record the page should reload with an
  error message: "Duplicate patron record?"
- Click "View existing record"
- A pop-up window should appear with patron information.
- Confirm that the window's footer looks correct.

- Locate a patron with an age-limited child category. In the sample data
  the "Juvenile" category works.
- View the patron detail page and click More -> Update child to adult
  patron.
- A pop-up window should appear with patron category selection options.
- Try submitting the form without making a selection. You should see a
  message about fields being required.
- Select a category and submit the form to confirm that the controls
  work correctly.

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>