Bug 35363: Update transfer order pop-up window with consistent footer markup
This patch updates the "Transfer order" pop-up window so that it has a
fixed footer consistent with other similar pop-up windows.

In order to make navigation between pages within the pop-up window
easier I've added a button in the footer which can take the user back to
the previous step, including any necessary parameters.

To test, apply the patch and go to Acquisitions.

- Locate a vendor with a basket, and a basket with one or more orders.
- On the basket detail page, click the "Transfer" link in the "Modify"
  column.
- In the pop-up window, test each step:
  - Vendor search
  - Vendor selection
  - Basket selection
- Confirm that at any step you can use the new button to go back to the
  previous step.

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>