Bug 35364: Update serials pop-up windows with consistent footer markup
This patch updates several serials module templates in order to style
submission and close buttons in a fixed footer at the bottom of pop-up
windows, with markup consistent with other pop-up windows.

Unrelated change: A couple of templates have had a "page-section" div
added where it was missing.

To test, apply the patch and go to Serials. For each affected template,
test that the window looks correct and the buttons in the footer work
correctly:

- New subscription ->
  - Search for a vendor:
    - Vendor search form
      - Vendor search results
  - Search for record
    - Record search form
    - Record search results
- Complete the process of creating a subscription if there are no other
  subscriptions already in your database.
  - From the subscription detail page, click the "Renew" button in the
    toolbar and test that window.
  - From the subscription detail page, click the "Create routing list"
    link in the left-hand sidebar menu.
    - Add one or more recipients and click "Save"
    - Click "Save and preview routing slip" and test the resulting
     pop-up.

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>