Bug 35407: Correct terminology: Show fewer collections
In the OPAC, the "Show more" and "Show less" links have aria-labels to
make their function clearer to screen readers. The collection and item
type controls need corrected terminology, and the links also need to
have their template filter corrected (html, not uri).

To test, apply the patch and perform a catalog search in the OPAC.

- Test that the "Show more" and "Show less" links in the facets sidebar
  work correctly to expand and collapse the lists of facets.
- It's not necessary to test with a screen reader. You can view the
  source to confirm the markup on the "Show more collections" and "Show
  more item types" links.

  Before:
  <a href="#" aria-label="%0A%20%20%20%20Show%20more%20collection%20codes%0A"><strong>Show more</strong></a>

  After:
  <a href="#" aria-label="Show more collections"><strong>Show more</strong></a>

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>
Signed-off-by: Victor Grousset/tuxayo <victor@tuxayo.net>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>