Bug 35556: Fix random selenium failure if category on second page
If 'Test category' if on the second page the test will fail to
locate the Edit button.

To make sure it is displayed on the screen we are selecting "All" in the
dropdown list.

This is done a couple of lines later in the script for the same purpose.

Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>
(cherry picked from commit 6f45fd45742b1fcea65bb7f626eadfedcbaeb941)
Signed-off-by: Fridolin Somers <fridolin.somers@biblibre.com>